Move toggle admin status to own file
